semphoni-sem-client/requirements.txt
flask
pyautogui
Pillow
python-dotenv
mss
websockets